.claude/skills/generate-image/requirements.txt
click
python-dotenv
Pillow
